+++ app/src/main/java/com/example/nicocalculator/mvp/view/CalculatorView.kt
@@ -1,0 +1,15 @@
+package com.example.nicocalculator.mvp.view
+
+import android.app.Activity
+import kotlinx.android.synthetic.main.activity_main.*
+
+class KeyView(private val activity: Activity) {
+
+    fun setTextResult(string: String) {
+        activity.text_results.text = string
+    }
+
+    fun setTextOperations(string: String) {
+        activity.text_operations.text = string
+    }
+}
